requirements.txt
Babel==2.4.0
Django==1.11.3
django-geojson==2.10.0
django-guardian==1.4.9
django-leaflet==0.22.0
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
jsonfield==2.0.2
phonenumberslite==8.6.0
psycopg2==2.7.1
pytz==2017.2
six==1.10.0
